+++ src/test/java/com/github/cc3002/finalreality/model/character/player/ThiefTest.java
@@ -1,8 +1,13 @@
 package com.github.cc3002.finalreality.model.character.player;
 
 import com.github.cc3002.finalreality.model.character.Enemy;
+import com.github.cc3002.finalreality.model.weapon.IWeapon;
+import com.github.cc3002.finalreality.model.weapon.KnifeWeapon;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.Arrays;
 
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
@@ -19,10 +24,11 @@ public class ThiefTest extends AbstractPlayerTest {
     /**
      * Setup method.
      * Creates a new thief named Zidane with 10 speed and links it to a turn queue.
+     * Create the main weapon
      */
     @BeforeEach
     void setUp() {
-        super.basicSetup();
+        super.basicSetup(new KnifeWeapon("mainWeapon", 8, 10));
         super.assignCharacter(new Thief(turns, THIEF_NAME));
     }
 
@@ -36,6 +42,21 @@ public class ThiefTest extends AbstractPlayerTest {
                 new WhiteMage(turns, THIEF_NAME),
                 new Enemy("Enemy", 10, turns));
         assertNotEquals(testCharacter, new WhiteMage(turns, THIEF_NAME));
+    }
+
+    /**
+     * Checks that the Thief character only equip compatible weapons.
+     */
+    @Test
+    void weaponEquipmentTest(){
+        ArrayList<IWeapon> supported = new ArrayList<>(Arrays.asList(
+                testSword, testKnife, testBow
+        ));
+        equipSupportedWeapons(supported);
+        ArrayList<IWeapon> unSupported = new ArrayList<>(Arrays.asList(
+                testAxe, testStaff
+        ));
+        equipUnsupportedWeapons(unSupported);
     }
 
 }