+++ src/test/java/com/github/cc3002/finalreality/model/character/player/WhiteMageTest.java
@@ -1,8 +1,13 @@
 package com.github.cc3002.finalreality.model.character.player;
 
 import com.github.cc3002.finalreality.model.character.Enemy;
+import com.github.cc3002.finalreality.model.weapon.IWeapon;
+import com.github.cc3002.finalreality.model.weapon.StaffWeapon;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.Arrays;
 
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
@@ -19,10 +24,11 @@ public class WhiteMageTest extends AbstractPlayerTest {
     /**
      * Setup method.
      * Creates a new white mage named Eiko with 10 speed and links it to a turn queue.
+     * Create the main weapon
      */
     @BeforeEach
     void setUp() {
-        super.basicSetup();
+        super.basicSetup(new StaffWeapon("mainWeapon", 3, 10));
         super.assignCharacter(new WhiteMage(turns, WHITE_MAGE_NAME));
     }
 
@@ -36,6 +42,21 @@ public class WhiteMageTest extends AbstractPlayerTest {
                 new Knight(turns, WHITE_MAGE_NAME),
                 new Enemy("Enemy", 10, turns));
         assertNotEquals(testCharacter, new BlackMage(turns, WHITE_MAGE_NAME));
+    }
+
+    /**
+     * Checks that the White Mage character only equip compatible weapons.
+     */
+    @Test
+    void weaponEquipmentTest(){
+        ArrayList<IWeapon> supported = new ArrayList<>(Arrays.asList(
+                testStaff
+        ));
+        equipSupportedWeapons(supported);
+        ArrayList<IWeapon> unSupported = new ArrayList<>(Arrays.asList(
+                testSword, testAxe, testBow, testKnife
+        ));
+        equipUnsupportedWeapons(unSupported);
     }
 
 }